backend/Deploy/API/requirements.txt
flask
pandas
redis
PyPDF2
nltk
torch
nltk
transformers
gunicorn
google-cloud-storage